python/requirements.txt
streamlit>=1.31
requests>=2.31
websocket-client>=1.6
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
pyserial>=3.5